+++ cucumber-junit-platform-engine/src/test/java/io/cucumber/junit/platform/engine/CucumberTestEngineTest.java
@@ -1,35 +1,81 @@
 package io.cucumber.junit.platform.engine;
 
-import org.assertj.core.api.Assertions;
-import org.assertj.core.api.Condition;
+import io.cucumber.core.logging.LogRecordListener;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.MethodSource;
+import org.junit.platform.commons.JUnitException;
+import org.junit.platform.commons.support.Resource;
 import org.junit.platform.engine.ConfigurationParameters;
 import org.junit.platform.engine.EngineDiscoveryRequest;
 import org.junit.platform.engine.EngineExecutionListener;
 import org.junit.platform.engine.ExecutionRequest;
 import org.junit.platform.engine.TestDescriptor;
-import org.junit.platform.engine.TestSource;
 import org.junit.platform.engine.UniqueId;
+import org.junit.platform.engine.discovery.DiscoverySelectors;
+import org.junit.platform.engine.discovery.FilePosition;
 import org.junit.platform.engine.support.descriptor.ClassSource;
+import org.junit.platform.engine.support.descriptor.ClasspathResourceSource;
+import org.junit.platform.engine.support.descriptor.FileSource;
 import org.junit.platform.testkit.engine.EngineTestKit;
 import org.junit.platform.testkit.engine.Event;
-import org.junit.platform.testkit.engine.EventConditions;
 
+import java.io.File;
+import java.net.URI;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.util.Arrays;
+import java.util.LinkedHashSet;
 import java.util.Optional;
+import java.util.Set;
+import java.util.function.Predicate;
+import java.util.logging.Level;
+import java.util.logging.LogRecord;
+import java.util.stream.Collectors;
 
 import static io.cucumber.junit.platform.engine.Constants.FEATURES_PROPERTY_NAME;
 import static io.cucumber.junit.platform.engine.Constants.FILTER_NAME_PROPERTY_NAME;
 import static io.cucumber.junit.platform.engine.Constants.FILTER_TAGS_PROPERTY_NAME;
-import static io.cucumber.junit.platform.engine.Constants.PLUGIN_PUBLISH_QUIET_PROPERTY_NAME;
+import static io.cucumber.junit.platform.engine.Constants.JUNIT_PLATFORM_LONG_NAMING_STRATEGY_EXAMPLE_NAME_PROPERTY_NAME;
+import static io.cucumber.junit.platform.engine.Constants.JUNIT_PLATFORM_NAMING_STRATEGY_PROPERTY_NAME;
+import static io.cucumber.junit.platform.engine.Constants.JUNIT_PLATFORM_SHORT_NAMING_STRATEGY_EXAMPLE_NAME_PROPERTY_NAME;
 import static io.cucumber.junit.platform.engine.CucumberEngineDescriptor.ENGINE_ID;
+import static io.cucumber.junit.platform.engine.CucumberEventConditions.emptySource;
+import static io.cucumber.junit.platform.engine.CucumberEventConditions.engine;
+import static io.cucumber.junit.platform.engine.CucumberEventConditions.example;
+import static io.cucumber.junit.platform.engine.CucumberEventConditions.examples;
+import static io.cucumber.junit.platform.engine.CucumberEventConditions.feature;
+import static io.cucumber.junit.platform.engine.CucumberEventConditions.prefix;
+import static io.cucumber.junit.platform.engine.CucumberEventConditions.rule;
+import static io.cucumber.junit.platform.engine.CucumberEventConditions.scenario;
+import static io.cucumber.junit.platform.engine.CucumberEventConditions.source;
+import static io.cucumber.junit.platform.engine.CucumberEventConditions.tags;
+import static java.util.Collections.emptySet;
+import static java.util.Collections.singleton;
+import static java.util.stream.Collectors.toList;
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.platform.engine.UniqueId.forEngine;
+import static org.junit.platform.engine.discovery.DiscoverySelectors.selectClass;
+import static org.junit.platform.engine.discovery.DiscoverySelectors.selectClasspathResource;
+import static org.junit.platform.engine.discovery.DiscoverySelectors.selectClasspathRoots;
+import static org.junit.platform.engine.discovery.DiscoverySelectors.selectDirectory;
 import static org.junit.platform.engine.discovery.DiscoverySelectors.selectFile;
+import static org.junit.platform.engine.discovery.DiscoverySelectors.selectPackage;
+import static org.junit.platform.engine.discovery.DiscoverySelectors.selectUniqueId;
+import static org.junit.platform.engine.discovery.DiscoverySelectors.selectUri;
+import static org.junit.platform.engine.discovery.PackageNameFilter.includePackageNames;
+import static org.junit.platform.engine.support.descriptor.FilePosition.from;
+import static org.junit.platform.testkit.engine.EventConditions.displayName;
 import static org.junit.platform.testkit.engine.EventConditions.event;
 import static org.junit.platform.testkit.engine.EventConditions.finishedSuccessfully;
 import static org.junit.platform.testkit.engine.EventConditions.skippedWithReason;
 import static org.junit.platform.testkit.engine.EventConditions.test;
 
+// TODO: Split out tests to multiple classes, but do use EngineTestKit everywhere
+@WithLogRecordListener
 class CucumberTestEngineTest {
 
     private final CucumberTestEngine engine = new CucumberTestEngine();
@@ -46,41 +92,415 @@ class CucumberTestEngineTest {
 
     @Test
     void createExecutionContext() {
+        // TODO: We don't need this test.
         EngineExecutionListener listener = new EmptyEngineExecutionListener();
         ConfigurationParameters configuration = new EmptyConfigurationParameters();
         EngineDiscoveryRequest discoveryRequest = new EmptyEngineDiscoveryRequest(configuration);
-        UniqueId id = UniqueId.forEngine(engine.getId());
+        UniqueId id = forEngine(engine.getId());
         TestDescriptor testDescriptor = engine.discover(discoveryRequest, id);
         ExecutionRequest execution = new ExecutionRequest(testDescriptor, listener, configuration);
         assertNotNull(engine.createExecutionContext(execution));
     }
 
     @Test
-    void selectAndExecuteNoScenario() {
+    void empty() {
         EngineTestKit.engine(ENGINE_ID)
-                .configurationParameter(PLUGIN_PUBLISH_QUIET_PROPERTY_NAME, "true")
                 .execute()
-                .testEvents()
+                .allEvents()
                 .assertThatEvents()
                 .haveExactly(0, event(test()));
     }
 
     @Test
-    void selectAndExecuteSingleScenario() {
+    void notCucumber() {
         EngineTestKit.engine(ENGINE_ID)
-                .configurationParameter(PLUGIN_PUBLISH_QUIET_PROPERTY_NAME, "true")
-                .selectors(selectFile("src/test/resources/io/cucumber/junit/platform/engine/single.feature"))
+                .selectors(selectUniqueId(forEngine("not-cucumber")))
                 .execute()
-                .testEvents()
+                .allEvents()
                 .assertThatEvents()
-                .haveExactly(2, event(test()))
-                .haveExactly(1, event(finishedSuccessfully()));
+                .haveExactly(0, event(test()));
     }
 
     @Test
-    void selectAndExecuteSingleScenarioThroughFeaturesProperty() {
+    void supportsClassSelector() {
         EngineTestKit.engine(ENGINE_ID)
-                .configurationParameter(PLUGIN_PUBLISH_QUIET_PROPERTY_NAME, "true")
+                .selectors(selectClass(RunCucumberTest.class))
+                .execute()
+                .containerEvents()
+                .assertEventsMatchLooselyInOrder(
+                    feature("disabled.feature"),
+                    feature("empty-scenario.feature"),
+                    feature("scenario-outline.feature"),
+                    feature("rule.feature"),
+                    feature("single.feature"),
+                    feature("with%20space.feature"));
+    }
+
+    @Test
+    void supportsClasspathResourceSelector() {
+        EngineTestKit.engine(ENGINE_ID)
+                .selectors(selectClasspathResource("io/cucumber/junit/platform/engine/single.feature"))
+                .execute()
+                .allEvents()
+                .assertThatEvents()
+                .haveExactly(1, event( //
+                    scenario("scenario:3", "A single scenario"), //
+                    finishedSuccessfully()));
+    }
+
+    @Test
+    void warnWhenResourceSelectorIsUsedToSelectAPackage(LogRecordListener logRecordListener) {
+        EngineTestKit.engine(ENGINE_ID)
+                .selectors(selectClasspathResource("io/cucumber/junit/platform/engine"))
+                .execute()
+                .allEvents()
+                .assertEventsMatchLooselyInOrder(
+                    feature("disabled.feature"),
+                    feature("empty-scenario.feature"),
+                    feature("scenario-outline.feature"),
+                    feature("rule.feature"),
+                    feature("single.feature"),
+                    feature("with%20space.feature"));
+
+        LogRecord warning = logRecordListener.getLogRecords()
+                .stream()
+                .filter(logRecord -> FeatureResolver.class.getName().equals(logRecord.getLoggerName()))
+                .filter(logRecord -> Level.WARNING.equals(logRecord.getLevel()))
+                .findFirst().get();
+
+        assertThat(warning.getMessage())
+                .isEqualTo(
+                    "The classpath resource selector 'io/cucumber/junit/platform/engine' should not be " +
+                            "used to select features in a package. Use the package selector with " +
+                            "'io.cucumber.junit.platform.engine' instead");
+    }
+
+    @Test
+    void classpathResourceSelectorThrowIfDuplicateResources() {
+        class TestResource implements Resource {
+
+            private final String name;
+            private final File source;
+
+            TestResource(String name, File source) {
+                this.name = name;
+                this.source = source;
+            }
+
+            @Override
+            public String getName() {
+                return name;
+            }
+
+            @Override
+            public URI getUri() {
+                return source.toURI();
+            }
+        }
+        Set<Resource> resources = new LinkedHashSet<>(Arrays.asList(
+            new TestResource("io/cucumber/junit/platform/engine/single.feature",
+                new File("src/test/resources/io/cucumber/junit/platform/engine/single.feature")),
+            new TestResource("io/cucumber/junit/platform/engine/single.feature",
+                new File("src/test/resources/io/cucumber/junit/platform/engine/single.feature")),
+            new TestResource("io/cucumber/junit/platform/engine/single.feature",
+                new File("src/test/resources/io/cucumber/junit/platform/engine/single.feature"))));
+
+        JUnitException exception = assertThrows(JUnitException.class, () -> EngineTestKit.engine(ENGINE_ID)
+                .selectors(selectClasspathResource(resources))
+                .execute()
+                .allEvents()
+                .assertThatEvents()
+                .haveExactly(1, event( //
+                    scenario("scenario:3", "A single scenario"), //
+                    finishedSuccessfully())));
+
+        assertThat(exception) //
+                .hasRootCauseInstanceOf(IllegalArgumentException.class) //
+                .hasRootCauseMessage( //
+                    "Found %s resources named %s on the classpath %s.", //
+                    resources.size(), //
+                    "io/cucumber/junit/platform/engine/single.feature", //
+                    resources.stream().map(Resource::getUri).collect(toList()));
+    }
+
+    @Test
+    void supportsClasspathResourceSelectorWithFilePosition() {
+        EngineTestKit.engine(ENGINE_ID)
+                .selectors(selectClasspathResource("io/cucumber/junit/platform/engine/rule.feature", //
+                    FilePosition.from(5)))
+                .execute()
+                .allEvents()
+                .assertThatEvents()
+                .haveExactly(2, event(scenario("scenario:5", "An example of this rule")));
+    }
+
+    @Test
+    void supportsMultipleClasspathResourceSelectors() {
+        EngineTestKit.engine(ENGINE_ID)
+                .selectors(
+                    selectClasspathResource("io/cucumber/junit/platform/engine/single.feature"),
+                    selectClasspathResource("io/cucumber/junit/platform/engine/scenario-outline.feature"))
+                .execute()
+                .allEvents()
+                .assertThatEvents()
+                .haveExactly(2, event(feature("single.feature", "A feature with a single scenario")))
+                .haveExactly(2, event(feature("scenario-outline.feature", "A feature with scenario outlines")));
+    }
+
+    @Test
+    void supportsClasspathResourceSelectorWithSpaceInResourceName() {
+        EngineTestKit.engine(ENGINE_ID)
+                .selectors(selectClasspathResource("io/cucumber/junit/platform/engine/with space.feature"))
+                .execute()
+                .allEvents()
+                .assertThatEvents()
+                .haveExactly(1, event(scenario(), finishedSuccessfully()));
+    }
+
+    @Test
+    void supportsClasspathRootSelector() {
+        Path classpathRoot = Paths.get("src/test/resources/");
+        EngineTestKit.engine(ENGINE_ID)
+                .selectors(selectClasspathRoots(singleton(classpathRoot)).get(0))
+                .execute()
+                .containerEvents()
+                .assertEventsMatchLooselyInOrder(
+                    feature("disabled.feature"),
+                    feature("empty-scenario.feature"),
+                    feature("scenario-outline.feature"),
+                    feature("rule.feature"),
+                    feature("single.feature"),
+                    feature("with%20space.feature"),
+                    feature("root.feature"));
+    }
+
+    @Test
+    void supportsDirectorySelector() {
+        EngineTestKit.engine(ENGINE_ID)
+                .selectors(selectDirectory("src/test/resources/io/cucumber/junit/platform/engine"))
+                .execute()
+                .containerEvents()
+                .assertEventsMatchLooselyInOrder(
+                    feature("disabled.feature"),
+                    feature("empty-scenario.feature"),
+                    feature("scenario-outline.feature"),
+                    feature("rule.feature"),
+                    feature("single.feature"),
+                    feature("with%20space.feature"));
+    }
+
+    @Test
+    void supportsFileSelector() {
+        EngineTestKit.engine(ENGINE_ID)
+                .selectors(selectFile("src/test/resources/io/cucumber/junit/platform/engine/single.feature"))
+                .execute()
+                .allEvents()
+                .assertThatEvents()
+                .haveExactly(1, event( //
+                    scenario("scenario:3", "A single scenario"), //
+                    finishedSuccessfully()));
+    }
+
+    @Test
+    void supportsFileSelectorWithFilePosition() {
+        EngineTestKit.engine(ENGINE_ID)
+                .selectors(selectFile("src/test/resources/io/cucumber/junit/platform/engine/rule.feature", //
+                    FilePosition.from(5)))
+                .execute()
+                .allEvents()
+                .assertThatEvents()
+                .haveExactly(1, event( //
+                    scenario("scenario:5", "An example of this rule"), //
+                    finishedSuccessfully()));
+    }
+
+    @Test
+    void supportsPackageSelector() {
+        EngineTestKit.engine(ENGINE_ID)
+                .selectors(selectPackage("io.cucumber.junit.platform.engine"))
+                .execute()
+                .containerEvents()
+                .assertEventsMatchLooselyInOrder(
+                    feature("disabled.feature"),
+                    feature("empty-scenario.feature"),
+                    feature("scenario-outline.feature"),
+                    feature("rule.feature"),
+                    feature("single.feature"),
+                    feature("with%20space.feature"));
+    }
+
+    @Test
+    void supportsUriSelector() {
+        File file = new File("src/test/resources/io/cucumber/junit/platform/engine/single.feature");
+        EngineTestKit.engine(ENGINE_ID)
+                .selectors(selectUri(file.toURI()))
+                .execute()
+                .allEvents()
+                .assertThatEvents()
+                .haveExactly(1, event( //
+                    scenario("scenario:3", "A single scenario"), //
+                    finishedSuccessfully()));
+    }
+
+    @Test
+    void supportsUriSelectorWithFilePosition() {
+        File file = new File("src/test/resources/io/cucumber/junit/platform/engine/rule.feature");
+        EngineTestKit.engine(ENGINE_ID)
+                .selectors(selectUri(file.toURI() + "?line=5"))
+                .execute()
+                .allEvents()
+                .assertThatEvents()
+                .haveExactly(1, event(scenario("scenario:5", "An example of this rule"), finishedSuccessfully()));
+    }
+
+    @ParameterizedTest
+    @MethodSource({
+            "supportsUniqueIdSelectorFromClasspathUri",
+            "supportsUniqueIdSelectorFromFileUri",
+            "supportsUniqueIdSelectorFromJarFileUri"
+    })
+    void supportsUniqueIdSelector(UniqueId selected) {
+        EngineTestKit.engine(ENGINE_ID)
+                .selectors(DiscoverySelectors.selectUniqueId(selected))
+                .execute()
+                .testEvents()
+                .assertThatEvents()
+                .haveAtLeastOne(event(prefix(selected), finishedSuccessfully()));
+    }
+
+    static Set<UniqueId> supportsUniqueIdSelectorFromClasspathUri() {
+        return EngineTestKit.engine(ENGINE_ID)
+                .selectors(selectPackage("io.cucumber.junit.platform.engine"))
+                .execute()
+                .allEvents()
+                .map(Event::getTestDescriptor)
+                .filter(Predicate.not(TestDescriptor::isRoot))
+                .map(TestDescriptor::getUniqueId)
+                .collect(Collectors.toSet());
+    }
+
+    static Set<UniqueId> supportsUniqueIdSelectorFromFileUri() {
+        return EngineTestKit.engine(ENGINE_ID)
+                .selectors(selectDirectory("src/test/resources/io/cucumber/junit/platform/engine"))
+                .execute()
+                .allEvents()
+                .map(Event::getTestDescriptor)
+                .filter(Predicate.not(TestDescriptor::isRoot))
+                .map(TestDescriptor::getUniqueId)
+                .collect(Collectors.toSet());
+    }
+
+    static Set<UniqueId> supportsUniqueIdSelectorFromJarFileUri() {
+        URI uri = new File("src/test/resources/feature.jar").toURI();
+        return EngineTestKit.engine(ENGINE_ID)
+                .selectors(selectUri(uri))
+                .execute()
+                .allEvents()
+                .map(Event::getTestDescriptor)
+                .filter(Predicate.not(TestDescriptor::isRoot))
+                .map(TestDescriptor::getUniqueId)
+                .collect(Collectors.toSet());
+    }
+
+    @Test
+    void supportsFilePositionFeature() {
+        EngineTestKit.engine(ENGINE_ID)
+                .selectors(
+                    selectFile("src/test/resources/io/cucumber/junit/platform/engine/scenario-outline.feature", //
+                        FilePosition.from(2)))
+                .execute()
+                .allEvents()
+                .assertThatEvents()
+                .haveExactly(2, event(feature("scenario-outline.feature", "A feature with scenario outlines")));
+    }
+
+    @Test
+    void supportsFilePositionScenario() {
+        EngineTestKit.engine(ENGINE_ID)
+                .selectors(
+                    selectFile("src/test/resources/io/cucumber/junit/platform/engine/scenario-outline.feature", //
+                        FilePosition.from(5)))
+                .execute()
+                .allEvents()
+                .assertThatEvents()
+                .haveExactly(1, event( //
+                    scenario("scenario:5", "A scenario"), //
+                    finishedSuccessfully()));
+    }
+
+    @Test
+    void supportsFilePositionScenarioOutline() {
+        EngineTestKit.engine(ENGINE_ID)
+                .selectors(
+                    selectFile("src/test/resources/io/cucumber/junit/platform/engine/scenario-outline.feature", //
+                        FilePosition.from(11)))
+                .execute()
+                .allEvents()
+                .assertThatEvents()
+                .haveExactly(1, event( //
+                    scenario("scenario:11", "A scenario outline"), //
+                    finishedSuccessfully()));
+    }
+
+    @Test
+    void supportsFilePositionExamples() {
+        EngineTestKit.engine(ENGINE_ID)
+                .selectors(
+                    selectFile("src/test/resources/io/cucumber/junit/platform/engine/scenario-outline.feature", //
+                        FilePosition.from(17)))
+                .execute()
+                .allEvents()
+                .assertThatEvents()
+                .haveExactly(1, event( //
+                    examples("examples:17", "With some text"), //
+                    finishedSuccessfully()));
+    }
+
+    @Test
+    void supportsFilePositionExample() {
+        EngineTestKit.engine(ENGINE_ID)
+                .selectors(
+                    selectFile("src/test/resources/io/cucumber/junit/platform/engine/scenario-outline.feature", //
+                        FilePosition.from(19)))
+                .execute()
+                .allEvents()
+                .assertThatEvents()
+                .haveExactly(1, event( //
+                    example("example:19", "Example #1.1"), //
+                    finishedSuccessfully()));
+    }
+
+    @Test
+    void supportsFilePositionRule() {
+        EngineTestKit.engine(ENGINE_ID)
+                .selectors(selectClasspathResource("io/cucumber/junit/platform/engine/rule.feature", //
+                    FilePosition.from(3)))
+                .execute()
+                .allEvents()
+                .assertThatEvents()
+                .haveExactly(2, event(rule("rule:3", "A rule")));
+    }
+
+    @Test
+    void executesFeaturesInUriOrderByDefault() {
+        EngineTestKit.engine(ENGINE_ID)
+                .selectors(selectPackage(""))
+                .execute()
+                .containerEvents()
+                .started()
+                .assertEventsMatchLooselyInOrder(
+                    feature("disabled.feature"),
+                    feature("empty-scenario.feature"),
+                    feature("scenario-outline.feature"),
+                    feature("rule.feature"),
+                    feature("single.feature"),
+                    feature("with%20space.feature"),
+                    feature("root.feature"));
+    }
+
+    @Test
+    void supportsFeaturesProperty() {
+        EngineTestKit.engine(ENGINE_ID)
                 .configurationParameter(FEATURES_PROPERTY_NAME,
                     "src/test/resources/io/cucumber/junit/platform/engine/single.feature")
                 .execute()
@@ -91,9 +511,32 @@ class CucumberTestEngineTest {
     }
 
     @Test
-    void selectAndExecuteSingleScenarioWithoutFeaturesProperty() {
+    @SuppressWarnings("OptionalGetWithoutIsPresent")
+    void supportsFeaturesPropertyWillIgnoreOtherSelectors(LogRecordListener logRecordListener) {
         EngineTestKit.engine(ENGINE_ID)
-                .configurationParameter(PLUGIN_PUBLISH_QUIET_PROPERTY_NAME, "true")
+                .configurationParameter(FEATURES_PROPERTY_NAME,
+                    "src/test/resources/io/cucumber/junit/platform/engine/single.feature")
+                .selectors(selectClasspathResource("io/cucumber/junit/platform/engine/rule.feature"))
+                .execute()
+                .allEvents()
+                .assertThatEvents()
+                .haveExactly(2, event(engine(source(ClassSource.from(CucumberTestEngine.class)))))
+                .haveExactly(1, event(test(finishedSuccessfully())));
+
+        LogRecord warning = logRecordListener.getLogRecords()
+                .stream()
+                .filter(logRecord -> FeaturesPropertyResolver.class.getName().equals(logRecord.getLoggerName()))
+                .filter(logRecord -> Level.WARNING.equals(logRecord.getLevel()))
+                .findFirst().get();
+
+        assertThat(warning.getMessage())
+                .startsWith(
+                    "Discovering tests using the cucumber.features property. Other discovery selectors are ignored!");
+    }
+
+    @Test
+    void onlySetsEngineSourceWhenFeaturesPropertyIsUsed() {
+        EngineTestKit.engine(ENGINE_ID)
                 .selectors(selectFile("src/test/resources/io/cucumber/junit/platform/engine/single.feature"))
                 .execute()
                 .allEvents()
@@ -105,7 +548,6 @@ class CucumberTestEngineTest {
     @Test
     void selectAndSkipDisabledScenarioByTags() {
         EngineTestKit.engine(ENGINE_ID)
-                .configurationParameter(PLUGIN_PUBLISH_QUIET_PROPERTY_NAME, "true")
                 .configurationParameter(FILTER_TAGS_PROPERTY_NAME, "@Integration and not @Disabled")
                 .selectors(selectFile("src/test/resources/io/cucumber/junit/platform/engine/single.feature"))
                 .execute()
@@ -119,28 +561,179 @@ class CucumberTestEngineTest {
     @Test
     void selectAndSkipDisabledScenarioByName() {
         EngineTestKit.engine(ENGINE_ID)
-                .configurationParameter(PLUGIN_PUBLISH_QUIET_PROPERTY_NAME, "true")
                 .configurationParameter(FILTER_NAME_PROPERTY_NAME, "^Nothing$")
                 .selectors(selectFile("src/test/resources/io/cucumber/junit/platform/engine/single.feature"))
                 .execute()
                 .testEvents()
                 .assertThatEvents()
-                .haveExactly(1, event(test()))
-                .haveExactly(1,
-                    event(skippedWithReason("'cucumber.filter.name=^Nothing$' did not match this scenario")));
+                .haveExactly(1, event(test(),
+                    event(skippedWithReason("'cucumber.filter.name=^Nothing$' did not match this scenario"))));
     }
 
-    private static Condition<Event> engine(Condition<Event> condition) {
-        return Assertions.allOf(EventConditions.engine(), condition);
+    @Test
+    void cucumberTagsAreConvertedToJunitTags() {
+        EngineTestKit.engine(ENGINE_ID)
+                .selectors(selectClasspathResource("io/cucumber/junit/platform/engine/scenario-outline.feature"))
+                .execute()
+                .allEvents()
+                .assertThatEvents()
+                .haveAtLeastOne(event(feature(), tags(emptySet())))
+                .haveAtLeastOne(
+                    event(scenario("scenario:5"), tags("FeatureTag", "ScenarioTag", "ResourceA", "ResourceAReadOnly")))
+                .haveAtLeastOne(event(scenario("scenario:11"), tags(emptySet())))
+                .haveAtLeastOne(event(examples("examples:17"), tags(emptySet())))
+                .haveAtLeastOne(event(example("example:19"), tags("FeatureTag", "ScenarioOutlineTag", "Example1Tag")));
     }
 
-    private static Condition<Event> source(TestSource testSource) {
-        return new Condition<>(event -> event.getTestDescriptor().getSource().filter(testSource::equals).isPresent(),
-            "test engine with test source '%s'", testSource);
+    @Test
+    void providesClasspathSourceWhenClasspathResourceIsSelected() {
+        String feature = "io/cucumber/junit/platform/engine/scenario-outline.feature";
+        EngineTestKit.engine(ENGINE_ID)
+                .selectors(selectClasspathResource(feature))
+                .execute()
+                .allEvents()
+                .assertThatEvents()
+                .haveAtLeastOne(event(feature(), source(ClasspathResourceSource.from(feature, from(2, 1)))))
+                .haveAtLeastOne(
+                    event(scenario("scenario:5"), source(ClasspathResourceSource.from(feature, from(5, 3)))))
+                .haveAtLeastOne(
+                    event(scenario("scenario:11"), source(ClasspathResourceSource.from(feature, from(11, 3)))))
+                .haveAtLeastOne(
+                    event(examples("examples:17"), source(ClasspathResourceSource.from(feature, from(17, 5)))))
+                .haveAtLeastOne(
+                    event(example("example:19"), source(ClasspathResourceSource.from(feature, from(19, 7)))));
     }
 
-    private static Condition<Event> emptySource() {
-        return new Condition<>(event -> !event.getTestDescriptor().getSource().isPresent(), "without a test source");
+    @Test
+    void providesFileSourceWhenFileIsSelected() {
+        File feature = new File("src/test/resources/io/cucumber/junit/platform/engine/scenario-outline.feature");
+        EngineTestKit.engine(ENGINE_ID)
+                .selectors(selectFile(feature))
+                .execute()
+                .allEvents()
+                .assertThatEvents()
+                .haveAtLeastOne(event(feature(), source(FileSource.from(feature, from(2, 1)))))
+                .haveAtLeastOne(event(scenario("scenario:5"), source(FileSource.from(feature, from(5, 3)))))
+                .haveAtLeastOne(event(scenario("scenario:11"), source(FileSource.from(feature, from(11, 3)))))
+                .haveAtLeastOne(event(examples("examples:17"), source(FileSource.from(feature, from(17, 5)))))
+                .haveAtLeastOne(event(example("example:19"), source(FileSource.from(feature, from(19, 7)))));
     }
 
+    @Test
+    void supportsPackageFilterForClasspathResources() {
+        Path classpathRoot = Paths.get("src/test/resources/");
+        EngineTestKit.engine(ENGINE_ID)
+                .selectors(selectClasspathRoots(singleton(classpathRoot)).get(0))
+                .filters(includePackageNames("io.cucumber.junit.platform"))
+                .execute()
+                .containerEvents()
+                .assertEventsMatchLooselyInOrder(
+                    feature("disabled.feature"),
+                    feature("empty-scenario.feature"),
+                    feature("scenario-outline.feature"),
+                    feature("rule.feature"),
+                    feature("single.feature"),
+                    feature("with%20space.feature"));
+    }
+
+    @Test
+    void supportsShortWithNumberAndPickleIfParameterizedNamingStrategy() {
+        EngineTestKit.engine(ENGINE_ID)
+                .configurationParameter(JUNIT_PLATFORM_NAMING_STRATEGY_PROPERTY_NAME, "short")
+                .configurationParameter(JUNIT_PLATFORM_SHORT_NAMING_STRATEGY_EXAMPLE_NAME_PROPERTY_NAME,
+                    "number-and-pickle-if-parameterized")
+                .selectors(
+                    selectClasspathResource("io/cucumber/junit/platform/engine/parameterized-scenario-outline.feature"))
+                .execute()
+                .allEvents()
+                .assertThatEvents()
+                .haveAtLeastOne(event(feature(), displayName("A feature with a parameterized scenario outline")))
+                .haveAtLeastOne(event(scenario(), displayName("A scenario full of <vegetable>s")))
+                .haveAtLeastOne(event(examples(), displayName("Of the Gherkin variety")))
+                .haveAtLeastOne(event(example(), displayName("Example #1.1: A scenario full of Cucumbers")));
+    }
+
+    @Test
+    void defaultsToShortWithNumberAndPickleIfParameterizedNamingStrategy() {
+        EngineTestKit.engine(ENGINE_ID)
+                .selectors(
+                    selectClasspathResource("io/cucumber/junit/platform/engine/parameterized-scenario-outline.feature"))
+                .execute()
+                .allEvents()
+                .assertThatEvents()
+                .haveAtLeastOne(event(feature(), displayName("A feature with a parameterized scenario outline")))
+                .haveAtLeastOne(event(scenario(), displayName("A scenario full of <vegetable>s")))
+                .haveAtLeastOne(event(examples(), displayName("Of the Gherkin variety")))
+                .haveAtLeastOne(event(example(), displayName("Example #1.1: A scenario full of Cucumbers")));
+    }
+
+    @Test
+    void supportsLongWithNumberNamingStrategy() {
+        EngineTestKit.engine(ENGINE_ID)
+                .configurationParameter(JUNIT_PLATFORM_NAMING_STRATEGY_PROPERTY_NAME, "long")
+                .configurationParameter(JUNIT_PLATFORM_LONG_NAMING_STRATEGY_EXAMPLE_NAME_PROPERTY_NAME, "number")
+                .selectors(
+                    selectClasspathResource("io/cucumber/junit/platform/engine/parameterized-scenario-outline.feature"))
+                .execute()
+                .allEvents()
+                .assertThatEvents()
+                .haveAtLeastOne(event(feature(), displayName("A feature with a parameterized scenario outline")))
+                .haveAtLeastOne(event(scenario(),
+                    displayName("A feature with a parameterized scenario outline - A scenario full of <vegetable>s")))
+                .haveAtLeastOne(event(examples(), displayName(
+                    "A feature with a parameterized scenario outline - A scenario full of <vegetable>s - Of the Gherkin variety")))
+                .haveAtLeastOne(event(example(), displayName(
+                    "A feature with a parameterized scenario outline - A scenario full of <vegetable>s - Of the Gherkin variety - Example #1.1")));
+    }
+
+    @Test
+    void supportsLongWithPickleNamingStrategy() {
+        EngineTestKit.engine(ENGINE_ID)
+                .configurationParameter(JUNIT_PLATFORM_NAMING_STRATEGY_PROPERTY_NAME, "long")
+                .configurationParameter(JUNIT_PLATFORM_LONG_NAMING_STRATEGY_EXAMPLE_NAME_PROPERTY_NAME, "pickle")
+                .selectors(
+                    selectClasspathResource("io/cucumber/junit/platform/engine/parameterized-scenario-outline.feature"))
+                .execute()
+                .allEvents()
+                .assertThatEvents()
+                .haveAtLeastOne(event(feature(), displayName("A feature with a parameterized scenario outline")))
+                .haveAtLeastOne(event(scenario(),
+                    displayName("A feature with a parameterized scenario outline - A scenario full of <vegetable>s")))
+                .haveAtLeastOne(event(examples(), displayName(
+                    "A feature with a parameterized scenario outline - A scenario full of <vegetable>s - Of the Gherkin variety")))
+                .haveAtLeastOne(event(example(), displayName(
+                    "A feature with a parameterized scenario outline - A scenario full of <vegetable>s - Of the Gherkin variety - A scenario full of Cucumbers")));
+    }
+
+    @Test
+    void supportsShortWithPickleNamingStrategy() {
+        EngineTestKit.engine(ENGINE_ID)
+                .configurationParameter(JUNIT_PLATFORM_NAMING_STRATEGY_PROPERTY_NAME, "short")
+                .configurationParameter(JUNIT_PLATFORM_SHORT_NAMING_STRATEGY_EXAMPLE_NAME_PROPERTY_NAME, "pickle")
+                .selectors(
+                    selectClasspathResource("io/cucumber/junit/platform/engine/parameterized-scenario-outline.feature"))
+                .execute()
+                .allEvents()
+                .assertThatEvents()
+                .haveAtLeastOne(event(feature(), displayName("A feature with a parameterized scenario outline")))
+                .haveAtLeastOne(event(scenario(), displayName("A scenario full of <vegetable>s")))
+                .haveAtLeastOne(event(examples(), displayName("Of the Gherkin variety")))
+                .haveAtLeastOne(event(example(), displayName("A scenario full of Cucumbers")));
+    }
+
+    @Test
+    void supportsShortWithNumberNamingStrategy() {
+        EngineTestKit.engine(ENGINE_ID)
+                .configurationParameter(JUNIT_PLATFORM_NAMING_STRATEGY_PROPERTY_NAME, "short")
+                .configurationParameter(JUNIT_PLATFORM_SHORT_NAMING_STRATEGY_EXAMPLE_NAME_PROPERTY_NAME, "number")
+                .selectors(
+                    selectClasspathResource("io/cucumber/junit/platform/engine/parameterized-scenario-outline.feature"))
+                .execute()
+                .allEvents()
+                .assertThatEvents()
+                .haveAtLeastOne(event(feature(), displayName("A feature with a parameterized scenario outline")))
+                .haveAtLeastOne(event(scenario(), displayName("A scenario full of <vegetable>s")))
+                .haveAtLeastOne(event(examples(), displayName("Of the Gherkin variety")))
+                .haveAtLeastOne(event(example(), displayName("Example #1.1")));
+    }
 }